B - Eröffnuungstext in AB wird auf 2. Seite noch einmal gebracht

+++ app/View/Elements/backend/SheetDelivery.ctp
@@ -22,7 +22,7 @@ foreach ($pages as $page => $carti) {
 <article class="module width_full sheet business noInput">	
 		<?php 
 			echo $this->element('backend/portlets/Cheet/header', array('pdf' => $pdf, 'page' => $pageCur, 'maxPage' => count($pages), 'logo' => true)); 
-			if($this->data['Confirmation']['order_date']) {			
+			if($this->data['Confirmation']['order_date'] && $pageCur < 2) {			
 				if($this->data['Confirmation']['pattern']) {
 					echo '<p class="offerText">'.sprintf(Configure::read('padcon.Lieferschein.header.pattern'),date('d.m.Y',strtotime($this->data['Confirmation']['order_date']))).'</p>';	
 				} else {
